+++ anzip/lib/src/model/extra_data_record.dart
@@ -1,6 +1,6 @@
 class ExtraDataRecord {
   const ExtraDataRecord(this.signature, this.data);
 
-  final int signature;
+  final List<int> signature;
   final List<int> data;
 }