+++ anzip/lib/src/model/zip_64_end_of_central_directory_locator.dart
@@ -13,7 +13,7 @@ class Zip64EndOfCentralDirectoryLocator implements ZipHeader {
   );
 
   @override
-  int get signature => kZip64endsig;
+  List<int> get signature => kZip64endsig;
 
   /// number of the disk with the start of the zip64 end of central directory
   final int numberOfDiskStartOfZip64EndOfCentralDirectoryRecord;
@@ -25,7 +25,7 @@ class Zip64EndOfCentralDirectoryLocator implements ZipHeader {
   final int totalNumberOfDisks;
 
   Future<void> write(BufferedSink sink) async {
-    await sink.writeUint32(signature, Endian.little);
+    await sink.writeFromBytes(signature);
     await sink.writeUint32(
         numberOfDiskStartOfZip64EndOfCentralDirectoryRecord, Endian.little);
     await sink.writeUint64(