+++ anzip/lib/src/model/zip_header.dart
@@ -1,3 +1,3 @@
 abstract interface class ZipHeader {
-  int get signature;
+  List<int> get signature;
 }